+++ jdkjson-api/src/main/java/org/jdkstack/jdkjson/api/pointer/Pointer.java
@@ -1,5 +1,6 @@
 package org.jdkstack.jdkjson.api.pointer;
 
 public interface Pointer {
-  //
+
+  Object value(Object json);
 }
